refactor(scorechain): move helpers into a util so the dto is types-only
